LPS-145112 - Move IconsUtil to kernal so that it can be accessed from portal-kernal, portal-web, and util-taglib
