+++ modules/apps/frontend-icons/frontend-icons-api/bnd.bnd
@@ -1,6 +1,4 @@
 Bundle-Name: Liferay Frontend Icons API
 Bundle-SymbolicName: com.liferay.frontend.icons.api
 Bundle-Version: 1.0.0
-Export-Package:\
-	com.liferay.frontend.icons.importer,\
-	com.liferay.frontend.icons.util
+Export-Package: com.liferay.frontend.icons.importer
