--- modules/apps/frontend-icons/frontend-icons-web/src/main/resources/com/liferay/frontend/icons/web/util/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
