+++ portal-kernel/src/com/liferay/portal/kernel/icons/IconsUtil.java
@@ -12,14 +12,17 @@
  * details.
  */
 
-package com.liferay.frontend.icons.util;
+package com.liferay.portal.kernel.icons;
 
+import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.StringBundler;
+
+// getPathThemeImages
 
 /**
  * @author Bryce Osterhaus
  */
-public class FrontendIconsUtil {
+public class IconsUtil {
 
 	public static String getBasePath() {
 		return _ICONS_BASE_PATH;
@@ -27,13 +30,19 @@ public class FrontendIconsUtil {
 
 	public static String getSpritemapPath(long siteId) {
 		return StringBundler.concat(
-			_ICONS_BASE_PATH, "/site/",
-			String.valueOf(siteId), ".svg");
+			_ICONS_BASE_PATH, "/site/", String.valueOf(siteId), ".svg");
 	}
 
 	public static String getSpritemapPath(String name) {
-		return StringBundler.concat(
-			_ICONS_BASE_PATH, "/pack/", name, ".svg");
+		return StringBundler.concat(_ICONS_BASE_PATH, "/pack/", name, ".svg");
+	}
+
+	public static String getSpritemapPath(ThemeDisplay themeDisplay) {
+		return getSpritemapPath(themeDisplay.getSiteGroupId());
+	}
+
+	public static String getSystemIconPackName() {
+		return _SYSTEM_ICON_PACK;
 	}
 
 	public static String getSystemSpritemapPath() {
@@ -41,6 +50,7 @@ public class FrontendIconsUtil {
 	}
 
 	private static final String _ICONS_BASE_PATH = "/o/icons";
+
 	private static final String _SYSTEM_ICON_PACK = "clay";
 
 }
