+++ portal-kernel/src/com/liferay/portal/kernel/icons/packageinfo
